+++ src/stores/show-state.ts
@@ -1,8 +1,8 @@
 import { atom } from 'nanostores'
 import { ShowState } from '../types';
 
-export const $show = atom<ShowState>(ShowState.BOTH)
+export const $showState = atom<ShowState>(ShowState.BOTH)
 
 export function setShow(newShowState: ShowState) {
-  $show.set(newShowState);
+  $showState.set(newShowState);
 }
